feat: add vector drawable resources for crop and pencil icons
